+++ Packages/ai.natml.vision.movenet-multipose/Changelog.md
@@ -1,2 +1,7 @@
+## 1.0.1
++ Added `MoveNetMultiposePredictor.Create` static method for creating the predictor.
++ Removed `MoveNetMultiposePredictor` public constructor.
++ Upgraded to NatML 1.1.2.
+
 ## 1.0.0
 + First release.